renamed 'TimeRequestWrapper'; changed to use 'dayOfArrival' too.
